backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-jose==3.3.0
httpx==0.27.0
groq==0.9.0
supabase==2.4.6
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
